+++ docs/changelog.rst
@@ -6,6 +6,9 @@ Changelog
 `Unreleased`_
 -------------
 
+`0.12.0`_ - 2021-07-08
+----------------------
+
 Fixed
 ~~~~~
 
@@ -171,7 +174,8 @@ Added
 
 - Initial public release
 
-.. _Unreleased: https://github.com/kiwicom/pytest-recording/compare/v0.11.0...HEAD
+.. _Unreleased: https://github.com/kiwicom/pytest-recording/compare/v0.12.0...HEAD
+.. _0.12.0: https://github.com/kiwicom/pytest-recording/compare/v0.11.0...v0.12.0
 .. _0.11.0: https://github.com/kiwicom/pytest-recording/compare/v0.10.0...v0.11.0
 .. _0.10.0: https://github.com/kiwicom/pytest-recording/compare/v0.9.0...v0.10.0
 .. _0.9.0: https://github.com/kiwicom/pytest-recording/compare/v0.8.1...v0.9.0